+++ eras/alonzo/test-suite/test/Tests.hs
@@ -13,10 +13,6 @@ import System.Environment (lookupEnv)
 import qualified Test.Cardano.Ledger.Alonzo.ChainTrace as ChainTrace
 import qualified Test.Cardano.Ledger.Alonzo.Golden as Golden
 import Test.Cardano.Ledger.Alonzo.ImpTest ()
-import qualified Test.Cardano.Ledger.Alonzo.Serialisation.Canonical as Canonical
-import qualified Test.Cardano.Ledger.Alonzo.Serialisation.Tripping as Tripping
-import qualified Test.Cardano.Ledger.Alonzo.Translation as Translation
-import qualified Test.Cardano.Ledger.Alonzo.TxInfo as TxInfo
 import qualified Test.Cardano.Ledger.Shelley.PropertyTests as Shelley
 import qualified Test.Cardano.Ledger.Shelley.Rules.AdaPreservation as AdaPreservation
 import qualified Test.Cardano.Ledger.Shelley.Rules.IncrementalStake as IncrementalStake
@@ -34,11 +30,7 @@ defaultTests =
   testGroup
     "Alonzo tests"
     [ AdaPreservation.tests @AlonzoEra 50
-    , Tripping.tests
-    , Translation.tests
-    , Canonical.tests
     , Golden.tests
-    , TxInfo.tests
     ]
 
 nightlyTests :: TestTree