+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Spec.hs
@@ -45,6 +45,7 @@ import Cardano.Ledger.Conway.Rules (
 import Cardano.Ledger.Conway.TxCert (ConwayTxCert)
 import Cardano.Ledger.Conway.TxInfo (ConwayContextError)
 import Cardano.Ledger.Plutus (Language (..))
+import Cardano.Ledger.Plutus.Language (SLanguage (..))
 import Cardano.Ledger.Shelley.API (ApplyTx)
 import Cardano.Ledger.Shelley.LedgerState (StashedAVVMAddresses)
 import Cardano.Ledger.Shelley.Rules (
@@ -53,9 +54,11 @@ import Cardano.Ledger.Shelley.Rules (
   ShelleyUtxowPredFailure,
  )
 import Control.State.Transition (STS (..))
+import Data.Proxy (Proxy (..))
 import Data.Typeable (Typeable)
 import qualified Test.Cardano.Ledger.Alonzo.Binary.CostModelsSpec as CostModelsSpec
 import qualified Test.Cardano.Ledger.Alonzo.Binary.TxWitsSpec as TxWitsSpec
+import qualified Test.Cardano.Ledger.Babbage.TxInfoSpec as BabbageTxInfo
 import Test.Cardano.Ledger.Common
 import qualified Test.Cardano.Ledger.Conway.Binary.Regression as Regression
 import qualified Test.Cardano.Ledger.Conway.BinarySpec as Binary
@@ -71,7 +74,8 @@ import Test.Cardano.Ledger.Core.JSON (roundTripJsonEraSpec)
 
 spec ::
   forall era.
-  ( EraPlutusTxInfo PlutusV2 era
+  ( EraPlutusTxInfo PlutusV1 era
+  , EraPlutusTxInfo PlutusV2 era
   , EraPlutusTxInfo PlutusV3 era
   , RuleListEra era
   , ConwayEraImp era
@@ -126,4 +130,9 @@ spec =
     describe "TxWits" $ do
       TxWitsSpec.spec @era
     Regression.spec @era
-    TxInfo.spec @era
+    describe "TxInfo" $ do
+      TxInfo.spec @era
+      BabbageTxInfo.spec (Proxy @era)
+      xdescribe "V2 tests for PlutusV3" $ do
+        -- TOOD: fix
+        BabbageTxInfo.txInfoSpecV2 (Proxy @era) SPlutusV3