+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Generator/Trace/Chain.hs
@@ -36,7 +36,6 @@ import Cardano.Ledger.Slot (
 import Cardano.Ledger.Val ((<->))
 import Cardano.Protocol.TPraos.API
 import Cardano.Protocol.TPraos.BHeader (
-  HashHeader (..),
   LastAppliedBlock (..),
   hashHeaderToNonce,
  )