+++ eras/conway/impl/CHANGELOG.md
@@ -81,6 +81,7 @@
 
 ### `testlib`
 
+* Added `Examples` with: `ledgerExamples`, `exampleConwayCerts`
 * Fix CDDL for `MultiAsset` in `TxOut` as well as the `Tx` mint field.
 * Add `mkConwayTestAccountState` and `conwayAccountsToUMap`
 * Rename `electCommittee` to `submitCommitteeElection` #5091